+++ src/pages/login.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import axios from "axios";
-import FormInput from "../components/layout/login/FormInput";
+import FormInput from "../components/layout/login";
 import logo from "../assets/odego_logo.png";
 
 const Login: React.FC = () => {
